WIP - moving constants in electron code
