+++ queuebot-client/main-process/main.ts
@@ -9,26 +9,17 @@ import { SongDownloader } from './downloader/song-downloader';
 import { SongDto } from '../../common';
 import * as cookie from 'cookie';
 import { environment } from './environment';
-
-// import {
-//   IPC_OPEN_TWITCH_LOGIN,
-//   IPC_SETTINGS_GET_VALUE,
-//   IPC_SETTINGS_SET_VALUE,
-//   LOGIN_URL,
-// } from './constants';
-
-export const IPC_OPEN_TWITCH_LOGIN = 'login.openTwitchLogin';
-export const IPC_SETTINGS_GET_VALUE = 'settings.getValue';
-export const IPC_SETTINGS_SET_VALUE = 'settings.setValue';
-export const IPC_SETTINGS_DELETE_VALUE = 'settings.deleteValue';
-export const LOGIN_URL = `${environment.queuebotApiBaseUrl}/auth/twitch`;
-export const IPC_SONG_DOWNLOADER_PROCESS_SONG = 'songDownloader.processSong';
-export const IPC_SONG_DOWNLOADER_PROCESS_SONG_PROGRESS =
-  'songDownloader.processSongProgress';
-
-export const IPC_PROTOCOL_HANDLER = 'login.protocolHandler';
-
-export const FILTERED_URLS = [`${environment.queuebotApiBaseUrl}/*`];
+import {
+  FILTERED_URLS,
+  IPC_OPEN_TWITCH_LOGIN,
+  IPC_PROTOCOL_HANDLER,
+  IPC_SETTINGS_DELETE_VALUE,
+  IPC_SETTINGS_GET_VALUE,
+  IPC_SETTINGS_SET_VALUE,
+  IPC_SONG_DOWNLOADER_PROCESS_SONG,
+  IPC_SONG_DOWNLOADER_PROCESS_SONG_PROGRESS,
+  LOGIN_URL,
+} from './constants';
 
 const settingsService = new SettingsStoreService(
   path.join(__dirname, 'settings.json'),
